requirements.txt
flask
mysql-connector-python
gunicorn
colorama
praw